requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
scikit-learn
mlflow
joblib
tensorflow
matplotlib
wordcloud
comet-ml
typing
opik
dvc
dvc-gs
fastapi
uvicorn 
python-multipart 
python-jose 
slowapi
redis
faiss-cpu
python-json-logger
prometheus-fastapi-instrumentator